菜单删除的修改

+++ src/main/java/com/sunkaisens/ibss/system/service/impl/MenuServiceImpl.java
@@ -1,9 +1,23 @@
 package com.sunkaisens.ibss.system.service.impl;
 
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.apache.commons.lang3.StringUtils;import org.apache.logging.log4j.util.StringBuilderFormattable;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.sun.tools.internal.xjc.generator.bean.ImplStructureStrategy.Result;
 import com.sunkaisens.ibss.common.domain.IBSSConstant;
 import com.sunkaisens.ibss.common.domain.Tree;
+import com.sunkaisens.ibss.common.exception.SysInnerException;
 import com.sunkaisens.ibss.common.utils.TreeUtil;
 import com.sunkaisens.ibss.system.dao.MenuMapper;
 import com.sunkaisens.ibss.system.domain.Menu;
@@ -11,19 +25,13 @@ import com.sunkaisens.ibss.system.manager.UserManager;
 import com.sunkaisens.ibss.system.service.MenuService;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.*;
 
 @Slf4j
 @Service("menuService")
 @Transactional(propagation = Propagation.SUPPORTS, readOnly = true, rollbackFor = Exception.class)
 public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu> implements MenuService {
-
+	
+	private String message;
     @Autowired
     private UserManager userManager;
 
@@ -96,17 +104,48 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu> implements Me
         this.userManager.loadUserPermissionRoleRedisCache(userIds);
     }
 
+    
+    /**
+     * 
+     * xsh 2019/8/6 删除的修改
+     */
     @Override
     @Transactional
-    public void deleteMeuns(String[] menuIds) throws Exception {
-        for (String menuId : menuIds) {
-            // 查找与这些菜单/按钮关联的用户
-            List<String> userIds = this.baseMapper.findUserIdsByMenuId(String.valueOf(menuId));
-            // 递归删除这些菜单/按钮
-            this.baseMapper.deleteMenus(menuId);
-            // 重新将这些用户的角色和权限缓存到 Redis中
-            this.userManager.loadUserPermissionRoleRedisCache(userIds);
-        }
+    public Map<String, Object> deleteMeuns(String[] menuIds) throws Exception {
+    	//实例化一个map；
+    	Map<String,Object> result = new HashMap<>();
+    	Menu menu = new Menu();//实例化一个menu
+  	       Long parentId=null;
+      for (String menuId : menuIds) {
+    	  menu=this.baseMapper.selectById(menuId);
+    	  parentId=menu.getParentId();
+    	  //先判断是不是父级菜单 如果不是可删除
+    	  if (parentId==0) {
+    		  result.put("flase", "0");
+    		  result.put("state","父级菜单不可删除" );
+    	  }else {
+    		 //查找与这些菜单/按钮关联的用户
+            try {
+				List<String> userIds = this.baseMapper.findUserIdsByMenuId(String.valueOf(menuId));
+				// 删除这些菜单/按钮
+				this.baseMapper.deleteMenus(menuId);
+				// 重新将这些用户的角色和权限缓存到 Redis中
+				this.userManager.loadUserPermissionRoleRedisCache(userIds);
+				result.put("success", "1");
+	    		result.put("state","删除成功" );
+			} catch (Exception e) {
+				// TODO 自动生成的 catch 块
+				result.put("flase", "0");
+	    		result.put("state","删除菜单/按钮失败" );
+				  message = "新增菜单/按钮失败";
+		          log.error(message, e);
+		          throw new SysInnerException(message);
+			}
+            
+    	  }
+    	  
+      }
+        return result;
     }
 
     private void buildTrees(List<Tree<Menu>> trees, List<Menu> menus, List<String> ids) {
